Fix up some type warnings.

+++ src/libmowgli/container/list.h
@@ -57,13 +57,13 @@ extern void mowgli_node_add(void *data, mowgli_node_t *n, mowgli_list_t *l);
 extern void mowgli_node_add_head(void *data, mowgli_node_t *n, mowgli_list_t *l);
 extern void mowgli_node_add_before(void *data, mowgli_node_t *n, mowgli_list_t *l, mowgli_node_t *before);
 extern void mowgli_node_add_after(void *data, mowgli_node_t *n, mowgli_list_t *l, mowgli_node_t *before);
-extern void mowgli_node_insert(void *data, mowgli_node_t *n, mowgli_list_t *l, int position);
-extern int mowgli_node_index(mowgli_node_t *n, mowgli_list_t *l);
+extern void mowgli_node_insert(void *data, mowgli_node_t *n, mowgli_list_t *l, size_t position);
+extern ssize_t mowgli_node_index(mowgli_node_t *n, mowgli_list_t *l);
 extern void mowgli_node_delete(mowgli_node_t *n, mowgli_list_t *l);
 extern mowgli_node_t *mowgli_node_find(void *data, mowgli_list_t *l);
 extern void mowgli_node_move(mowgli_node_t *m, mowgli_list_t *oldlist, mowgli_list_t *newlist);
-extern mowgli_node_t *mowgli_node_nth(mowgli_list_t *l, int pos);
-extern void *mowgli_node_nth_data(mowgli_list_t *l, int pos);
+extern mowgli_node_t *mowgli_node_nth(mowgli_list_t *l, size_t pos);
+extern void *mowgli_node_nth_data(mowgli_list_t *l, size_t pos);
 
 typedef int (*mowgli_list_comparator_t)(mowgli_node_t *n, mowgli_node_t *n2, void *opaque);
